completed lab2
