+++ lab2/bears.csv
@@ -1,0 +1,55 @@
+absolute_path;relative_path
+C:\Users\admin\Desktop\lab2\images\range_1\000001.jpg;range_1\000001.jpg
+C:\Users\admin\Desktop\lab2\images\range_1\000002.jpg;range_1\000002.jpg
+C:\Users\admin\Desktop\lab2\images\range_1\000003.jpg;range_1\000003.jpg
+C:\Users\admin\Desktop\lab2\images\range_1\000004.jpg;range_1\000004.jpg
+C:\Users\admin\Desktop\lab2\images\range_1\000005.jpg;range_1\000005.jpg
+C:\Users\admin\Desktop\lab2\images\range_1\000006.jpg;range_1\000006.jpg
+C:\Users\admin\Desktop\lab2\images\range_1\000007.jpg;range_1\000007.jpg
+C:\Users\admin\Desktop\lab2\images\range_1\000008.jpg;range_1\000008.jpg
+C:\Users\admin\Desktop\lab2\images\range_1\000009.jpg;range_1\000009.jpg
+C:\Users\admin\Desktop\lab2\images\range_1\000010.jpg;range_1\000010.jpg
+C:\Users\admin\Desktop\lab2\images\range_1\000011.jpg;range_1\000011.jpg
+C:\Users\admin\Desktop\lab2\images\range_1\000012.jpg;range_1\000012.jpg
+C:\Users\admin\Desktop\lab2\images\range_1\000013.jpg;range_1\000013.jpg
+C:\Users\admin\Desktop\lab2\images\range_1\000014.jpg;range_1\000014.jpg
+C:\Users\admin\Desktop\lab2\images\range_1\000015.jpg;range_1\000015.jpg
+C:\Users\admin\Desktop\lab2\images\range_1\000016.jpg;range_1\000016.jpg
+C:\Users\admin\Desktop\lab2\images\range_1\000017.jpg;range_1\000017.jpg
+C:\Users\admin\Desktop\lab2\images\range_1\000018.jpg;range_1\000018.jpg
+C:\Users\admin\Desktop\lab2\images\range_1\000019.png;range_1\000019.png
+C:\Users\admin\Desktop\lab2\images\range_1\000020.jpg;range_1\000020.jpg
+C:\Users\admin\Desktop\lab2\images\range_1\000021.jpg;range_1\000021.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000001.jpg;range_2\000001.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000002.jpg;range_2\000002.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000003.jpg;range_2\000003.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000004.jpg;range_2\000004.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000005.jpg;range_2\000005.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000006.jpg;range_2\000006.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000007.jpg;range_2\000007.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000008.jpg;range_2\000008.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000009.jpg;range_2\000009.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000010.jpg;range_2\000010.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000011.jpg;range_2\000011.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000012.jpg;range_2\000012.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000013.jpg;range_2\000013.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000014.jpg;range_2\000014.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000015.jpg;range_2\000015.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000016.jpg;range_2\000016.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000017.jpg;range_2\000017.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000018.jpg;range_2\000018.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000019.jpg;range_2\000019.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000020.jpg;range_2\000020.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000021.jpg;range_2\000021.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000022.jpg;range_2\000022.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000023.jpg;range_2\000023.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000024.png;range_2\000024.png
+C:\Users\admin\Desktop\lab2\images\range_2\000025.jpg;range_2\000025.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000026.jpg;range_2\000026.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000027.jpg;range_2\000027.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000028.jpg;range_2\000028.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000029.jpg;range_2\000029.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000030.jpg;range_2\000030.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000031.jpg;range_2\000031.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000032.jpg;range_2\000032.jpg
+C:\Users\admin\Desktop\lab2\images\range_2\000033.jpg;range_2\000033.jpg
